requirements.txt
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
pymongo==4.3.3
django-cors-headers==3.14.0
python-decouple==3.8
setuptools==69.5.1